// src/components/shared-types.tsx
export type WeeklyObject = {
    id: string
    weeklyName: string
    weeklyPrice: number
}

export type ProductObject = {
    id: string
    name: string
    price: number
    imageUrl: string
}

export type CharacterType = {
    id: string
    characterClass: string
    characterLevel: number
    characterName: string
}

export type WeeklyType = {
    id: string
    weeklyName: string
    weeklyPrice: number
}

export type BossCardObject = {
    id: string
    name: string;
    price: number;
    imageUrl: string;
}